requirements.txt
# Debug
django-debug-toolbar==0.11.0
django-extensions
Werkzeug
django-devserver
sqlparse

# Core
Django==1.9
ipython
ipdb
django-ses
django-braces
